Makes call to DBus asynchronous
Adds a LifereaWebKitImpl type to receive the DBus PageCreated signal
and emit a signal connected with all created LifereaWebView widgets.

Adds a LifereaWebView type inheriting from WebKitWebView to keep a
pointer to the DBus connection that emitted its page id.

The menu is done with actions on the LifereaWebView widget. This
removes the need for hasSelection and copySelection in the interface
with LifereaHtmlView.`webkit_web_view_can_execute_edit_command` is
used to enable/disable the corresponding action.

`scrollPagedown` is only used for headlines skimming.
A callback in LifereaWebView makes the call to go to next unread item
if scrolling scrolling reached the end of the page. `scrollPagedown` no
longer need to return anything.